moonData.py basic implementation
moonData.py uses Selenium and a headless (no-window) chrome browser to query mooncalc.org for the 10 closest overhead (90° altitude) approaches of the moon given coordinates in latitude and longitude.

The information queries can be sped up by skipping a number of days based on latitude.

I would like city options so that you do not have to know your coordinates.  I would also like to add a check for the coordinates or the closest city with the moon overhead at a given date and time.